refactor: move wildcard expansion state
